+++ src/main/java/com/jagrosh/jmusicbot/gui/ConsolePanel.java
@@ -30,7 +30,9 @@ public class ConsolePanel extends JPanel {
         text.setLineWrap(true);
         text.setWrapStyleWord(true);
         text.setEditable(false);
-        PrintStream con = new PrintStream(new TextAreaOutputStream(text));
+        
+        // Redirect System.out and System.err to the GUI console
+        PrintStream con = new PrintStream(new TextAreaOutputStream(text), true);
         System.setOut(con);
         System.setErr(con);